requirements.txt
streamlit==1.28.2
python-dotenv
openai
pandas
openpyxl
plotly==5.5.0
numpy==1.22.4
scipy